Moved files, updated refs and minor fixes
