+++ ui/src/routes/info/beta.svelte
@@ -1,5 +1,14 @@
 <script>
-	import Beta from 'app/components/content/Beta.svelte';
+	import Beta from 'app/components/content/info/Beta.svelte';
+	import {toolName} from 'app/config';
 </script>
+
+<svelte:head>
+	<title>Beta release - {toolName}</title>
+	<meta
+		content='{toolName}: Beta release'
+		name='description'
+	>
+</svelte:head>
 
 <Beta />
