+++ ui/src/routes/info/privacy.svelte
@@ -1,5 +1,6 @@
 <script>
-	import Privacy from 'app/components/content/Privacy.svelte';
+	import Privacy from 'app/components/content/info/Privacy.svelte';
+	import {toolName} from 'app/config';
 </script>
 
 <svelte:head>
